+++ R/specan.R
@@ -197,7 +197,7 @@ specan <- function(X, bp = c(0,22), wl = 512, wl.freq = NULL, threshold = 15,
   spFUN <- function(i, X, bp, wl, threshold) { 
     r <- tuneR::readWave(as.character(X$sound.files[i]), from = X$start[i], to = X$end[i], units = "seconds") 
     
-    if(bp[1] == "frange") b <- c(X$low.freq[i], X$high.freq[i])
+    if(bp[1] == "frange") b <- c(X$low.freq[i], X$high.freq[i]) else b <- bp
 
      #in case bp its higher than can be due to sampling rate
     if(b[2] > ceiling(r@samp.rate/2000) - 1) b[2] <- ceiling(r@samp.rate/2000) - 1